Housekeeping: Reorganize plugin modules
