+++ katana-mngr/katana/shared_utils/nfvoUtils/osmUtils.py
