+++ katana-mngr/katana/shared_utils/vimUtils/opennebulaUtils.py
