+++ katana-mngr/katana/shared_utils/vimUtils/openstackUtils.py
