+++ katana-nbi/katana/shared_utils/nfvoUtils/osmUtils.py
