+++ katana-nbi/katana/shared_utils/vimUtils/opennebulaUtils.py
