+++ katana-nbi/katana/shared_utils/vimUtils/openstackUtils.py
