+++ shared_utils/nfvoUtils/osmUtils.py
